phase-3-todo-ai-chatbot/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
openai
requests
openai-agents
python-jose[cryptography]
passlib[bcrypt]
python-multipart